refactor(release): unify GitHub publication
